google_drive/requirements.txt
Package                  Version
------------------------ ---------
cachetools               4.2.2
certifi                  2021.5.30
chardet                  4.0.0
google-api-core          1.30.0
google-api-python-client 2.9.0
google-auth              1.31.0
google-auth-httplib2     0.1.0
google-auth-oauthlib     0.4.4
googleapis-common-protos 1.53.0
httplib2                 0.19.1
idna                     2.10
oauth2client             4.1.3
oauthlib                 3.1.1
packaging                20.9
pip                      21.1.2
protobuf                 3.17.3
pyasn1                   0.4.8
pyasn1-modules           0.2.8
pyparsing                2.4.7
pytz                     2021.1
requests                 2.25.1
requests-oauthlib        1.3.0
rsa                      4.7.2
setuptools               57.0.0
six                      1.16.0
uritemplate              3.0.1
urllib3                  1.26.5
wheel                    0.36.2
